refactor(cli): extract runtime recommend handler
