+++ application/src/Nexo.CLI/Commands/RuntimeCommand.cs
@@ -13,8 +13,11 @@ namespace Nexo.CLI.Commands;
 /// </summary>
 public sealed class RuntimeCommand : Command
 {
+    private readonly RecommendHandler _recommendHandler;
+
     public RuntimeCommand() : base("runtime", "Adaptive runtime control plane commands.")
     {
+        _recommendHandler = new RecommendHandler();
         ConfigureExecuteCommand();
         ConfigurePlanCommand();
         ConfigureEvaluateCommand();
@@ -653,29 +656,7 @@ public sealed class RuntimeCommand : Command
         string goal,
         string repoRoot,
         int historyWindow,
-        bool json)
-    {
-        if (string.IsNullOrWhiteSpace(goal))
-        {
-            RuntimeOutputWriter.WriteRecommendResult(new RuntimeRecommendResult(false, "Goal is required."), json);
-            return Task.FromResult(1);
-        }
-
-        var fullRepoRoot = Path.GetFullPath(string.IsNullOrWhiteSpace(repoRoot) ? Environment.CurrentDirectory : repoRoot);
-        if (!Directory.Exists(fullRepoRoot))
-        {
-            RuntimeOutputWriter.WriteRecommendResult(new RuntimeRecommendResult(false, $"Repo root not found: {fullRepoRoot}"), json);
-            return Task.FromResult(1);
-        }
-
-        var history = AdaptiveRuntimeExecutionHistoryStore.ReadRecent(fullRepoRoot, Math.Max(1, historyWindow));
-        var recommendation = AdaptiveRuntimePolicyAdvisor.RecommendQaPolicy(goal, history);
-        var result = recommendation == null
-            ? new RuntimeRecommendResult(true, "No recommendation from history.", Policy: null, Reason: null)
-            : new RuntimeRecommendResult(true, "Recommendation computed from history.", recommendation.QaPolicy, recommendation.Reason);
-        RuntimeOutputWriter.WriteRecommendResult(result, json);
-        return Task.FromResult(0);
-    }
+        bool json) => _recommendHandler.ExecuteAsync(goal, repoRoot, historyWindow, json);
 
     internal Task<int> ExecuteGateAsync(
         string repoRoot,
